timeseries_api/requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio ==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
Bottleneck==1.3.7
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
dotenv==0.9.9
fastapi==0.116.1
Flask==3.1.1
uvicorn==0.35.0
gunicorn==21.2.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mkl-service==2.5.2
mkl_fft==1.3.8 
mkl-random==1.2.8
numexpr==2.8.4
numpy==1.26.4
osmnx==2.0.5
pandas==2.3.1
psycopg2-binary==2.9.10
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
pyvista==0.45.3
rasterio==1.4.3
requests==2.32.4
six==1.16.0
scikit-image==0.25.2
sniffio==1.3.1
starlette==0.47.2
PyYAML==6.0.2
typing-inspection==0.4.1
typing-extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
Werkzeug==3.1.3
